Rose-specific dependency decoupling part 1: modules
